Combo Box Catalogues et sous-catalogues de la Liste Articles OK

+++ SoftCaisse/Repositories/FBilletageRepository.cs
@@ -19,16 +19,22 @@ namespace SoftCaisse.Repositories
         public bool deleteRow(int cbMarq)
         {
             F_BILLETPIECE billet = _context.F_BILLETPIECE.FirstOrDefault(u=>u.cbMarq==cbMarq);
-            _context.F_BILLETPIECE.Remove(billet);
-            _context.SaveChanges();
+            if (billet != null)
+            {
+                _context.F_BILLETPIECE.Remove(billet);
+                _context.SaveChanges();
+            }
             return true;
         }
         public bool update(int cbMarq,decimal? valeur,string Intitule)
         {
             F_BILLETPIECE billet = _context.F_BILLETPIECE.FirstOrDefault(u=>u.cbMarq==cbMarq);
-            billet.BI_Valeur = valeur;
-            billet.BI_Intitule = Intitule;
-            _context.SaveChanges();
+            if (billet != null)
+            {
+                billet.BI_Valeur = valeur;
+                billet.BI_Intitule = Intitule;
+                _context.SaveChanges();
+            }
             return true;
         }
         public bool insert(int cbMarq, decimal? valeur, string Intitule,short? devise)
